Now actually resumes

+++ Emby.Mobile.Universal/Controls/MediaPlayers/BackgroundAudioPlayer.cs
@@ -276,7 +276,7 @@ namespace Emby.Mobile.Universal.Controls.MediaPlayers
             }
         }
 
-        public async Task Play(PlaylistItem item, double position = 0)
+        public async Task Play(PlaylistItem item, long position = 0)
         {
             _playlist = new List<PlaylistItem> { item };
             if (position > 0)
@@ -286,7 +286,7 @@ namespace Emby.Mobile.Universal.Controls.MediaPlayers
                 SendListToBackgroundPlayer(new List<TrackModel> { await GetTrackModel(item) }, true);
         }
 
-        public async Task Play(List<PlaylistItem> items, double position = 0d, int? startingItem = null)
+        public async Task Play(List<PlaylistItem> items, long position = 0, int? startingItem = null)
         {
             _playlist = items;
             var list = new List<TrackModel>();
